refactor: quantityfunction
