+++ src/contexts/cart-context.jsx
@@ -1,35 +1,23 @@
 import axios from "axios";
 import { createContext, useContext, useEffect, useReducer } from "react";
 import { useNavigate } from "react-router-dom";
+import { cartReducer } from "../reducers/cartReducer";
 import { useAuth } from "./auth-context";
 
 const CartContext = createContext(null);
 
 const useCart = () => useContext(CartContext);
 
-
 const initialState = {
   cartlist: [], //we pass an empty array as the initial state as initially the cart would be empty
 };
 
-const cartReducer = (state, action) => {
-  switch (action.type) {
-    case "ADD_TO_CART": {
-      return { ...state, cartlist: action.payload }; 
-    }
-    default:
-      return state;
-  }
-};
-
 const CartProvider = ({ children }) => {
-
   const navigate = useNavigate();
   const {
     auth: { Authenticated, token },
   } = useAuth(); // a function that returns something should be inside a functional component
 
-  // console.log("Hi")
   const [state, dispatch] = useReducer(cartReducer, initialState);
 
   //we run a useeffect here to get pre-saved cart items when the user logs in again
@@ -45,17 +33,14 @@ const CartProvider = ({ children }) => {
             },
           });
 
-          // console.log(response);
-
           dispatch({ type: "ADD_TO_CART", payload: response.data.cart });
         } catch (errors) {
           console.log(errors);
         }
 
         //we don't need anything in else here because if the user is not logged in then he won't be able to access cart in the first place.
-    
+      }
     }
-  }
 
     getCartItems();
   }, [token]); //calling this only when the token changes
