+++ src/contexts/wishlist-context.jsx
@@ -6,6 +6,7 @@ import {
   useEffect,
   useState,
 } from "react";
+import { wishListReducer } from "../reducers/wishlistReducer";
 import { useAuth } from "./auth-context";
 
 const wishlistContext = createContext(null);
@@ -24,22 +25,11 @@ const WishlistProvider = ({ children }) => {
   } = useAuth();
   const [errors, setErrors] = useState("");
 
-  const wishListReducer = (state, action) => {
-    // console.log(action);
-
-    switch (action.type) {
-      case "MOVE_TO_WISHLIST":
-        return { ...state, wishlist: action.payload }; // we can do this using useState as well as there's just one action type
-    }
-  };
-
   //reducer to fire the item from product to wishlist
 
   const [state, dispatch] = useReducer(wishListReducer, initialState);
 
   //This is used for us to fetch data from the backend to show data on the Wishlist page. We put this in useffect as we want it to run after the render and whenever the user logs in because it's a private route.
-
-  // console.log(Authenticated);
 
   useEffect(() => {
     async function getdataWishlist() {
@@ -53,8 +43,6 @@ const WishlistProvider = ({ children }) => {
           });
           setErrors("");
 
-          // console.log(response);
-
           //firing our payload
 
           dispatch({
@@ -64,8 +52,8 @@ const WishlistProvider = ({ children }) => {
         } catch (errors) {
           console.log(errors);
         }
-          //we don't need anything in else here because if the user is not logged in then he won't be able to access cart in the first place.
-      } 
+        //we don't need anything in else here because if the user is not logged in then he won't be able to access cart in the first place.
+      }
     }
 
     getdataWishlist(); //IIFE
